snapcast: fix circular import
